+++ ErsatzTV.Core/Interfaces/Streaming/IGraphicsElement.cs
@@ -4,6 +4,8 @@ namespace ErsatzTV.Core.Interfaces.Streaming;
 
 public interface IGraphicsElement
 {
+    bool IsFailed { get; set; }
+
     Task InitializeAsync(Resolution frameSize, int frameRate, CancellationToken cancellationToken);
 
     void Draw(object context, TimeSpan timestamp);