+++ ErsatzTV.Infrastructure/Streaming/WatermarkElement.cs
@@ -1,7 +1,9 @@
+using System.Globalization;
 using ErsatzTV.Core.Domain;
 using ErsatzTV.Core.FFmpeg;
 using ErsatzTV.Core.Interfaces.Streaming;
 using ErsatzTV.FFmpeg.State;
+using NCalc;
 using SixLabors.ImageSharp;
 using SixLabors.ImageSharp.Formats.Gif;
 using SixLabors.ImageSharp.Processing;
@@ -16,6 +18,7 @@ public class WatermarkElement : IGraphicsElement, IDisposable
     private readonly List<Image> _scaledFrames = [];
     private readonly List<double> _frameDelays = [];
 
+    private Expression _expression;
     private double _animatedDurationSeconds;
     private Image _sourceImage;
     private Point _location;
@@ -36,8 +39,20 @@ public class WatermarkElement : IGraphicsElement, IDisposable
 
     public bool IsValid => _imagePath != null && _watermark != null;
 
+    public bool IsFailed { get; set; }
+
     public async Task InitializeAsync(Resolution frameSize, int frameRate, CancellationToken cancellationToken)
     {
+        if (!string.IsNullOrWhiteSpace(_watermark.OpacityExpression))
+        {
+            _expression = new Expression(_watermark.OpacityExpression);
+        }
+        else
+        {
+            float opacity = _watermark.Opacity / 100.0f;
+            _expression = new Expression(opacity.ToString(CultureInfo.InvariantCulture));
+        }
+
         bool isRemoteUri = Uri.TryCreate(_imagePath, UriKind.Absolute, out var uriResult)
                            && (uriResult.Scheme == Uri.UriSchemeHttp || uriResult.Scheme == Uri.UriSchemeHttps);
 
@@ -73,13 +88,12 @@ public class WatermarkElement : IGraphicsElement, IDisposable
             horizontalMargin,
             verticalMargin);
 
-        float opacity = _watermark.Opacity / 100.0f;
         _animatedDurationSeconds = 0;
 
         for (int i = 0; i < _sourceImage.Frames.Count; i++)
         {
             var frame = _sourceImage.Frames.CloneFrame(i);
-            frame.Mutate(ctx => ctx.Resize(scaledWidth, scaledHeight).Opacity(opacity));
+            frame.Mutate(ctx => ctx.Resize(scaledWidth, scaledHeight));
             _scaledFrames.Add(frame);
 
             var frameDelay = _sourceImage.Frames[i].Metadata.GetFormatMetadata(GifFormat.Instance).FrameDelay / 100.0;
@@ -95,10 +109,16 @@ public class WatermarkElement : IGraphicsElement, IDisposable
             return;
         }
 
-        Image frameForTimestamp = GetFrameForTimestamp(timestamp);
+        _expression.Parameters["content_seconds"] = timestamp.TotalSeconds;
+        object expressionResult = _expression.Evaluate();
+        float opacity = Convert.ToSingle(expressionResult, CultureInfo.InvariantCulture);
+        if (opacity == 0)
+        {
+            return;
+        }
 
-        // scaled frames already have opacity set
-        imageProcessingContext.DrawImage(frameForTimestamp, _location, 1f);
+        Image frameForTimestamp = GetFrameForTimestamp(timestamp);
+        imageProcessingContext.DrawImage(frameForTimestamp, _location, opacity);
     }
 
     private Image GetFrameForTimestamp(TimeSpan timestamp)